🔨 fix(server): warning comparation between differentes types of variables

+++ server/src/queue/enqueue.c
@@ -13,7 +13,7 @@ void print_command_enqueue(client_t* client, const char* command, int freq)
     int action_time = 0;
     command_queue_t* queue = &client->player->command_queue;
 
-    for (int i = 0; i < PLAYER_COMMANDS_SIZE; i++) {
+    for (size_t i = 0; i < PLAYER_COMMANDS_SIZE; i++) {
         if (!strcmp(command, PLAYER_COMMANDS[i].name)) {
             action_time = PLAYER_COMMANDS[i].action_time;
             break;
@@ -50,7 +50,7 @@ void enqueue_command(client_t* client, const char* command, int freq)
         queue->rear = (queue->rear + 1) % MAX_COMMANDS_PER_CLIENT;
     }
 
-    for (int i = 0; i < PLAYER_COMMANDS_SIZE; i++) {
+    for (size_t i = 0; i < PLAYER_COMMANDS_SIZE; i++) {
         if (!strcmp(command, PLAYER_COMMANDS[i].name)) {
             action_time = PLAYER_COMMANDS[i].action_time;
             break;
